Refactor: 관심사 분리 - 채팅 페이지 내 auth 로직
