+++ src/pages/ChattingPage/index.tsx
@@ -1,4 +1,4 @@
-import { API_BASE_URL, API_SUB_URLS, API_SUB_URLS_V2 } from '@/constants/apiConfig';
+import { API_BASE_URL, API_SUB_URLS_V2 } from '@/constants/apiConfig';
 import ChatbotSystemMessage from '@/features/chatbot/components/ChatbotSystemMessage';
 import useInput from '@/shared/hooks/useInput';
 import BottomNav from '@/shared/layout/BottomNav';
@@ -9,6 +9,8 @@ import MDEditor from '@uiw/react-md-editor';
 import useGetSessionHistory from '@/features/session/hooks/useGetSessionHistory';
 import { IMessage } from '@/features/comment/types/message';
 import { handleSSEStream } from '@/features/chatbot/utils/handleSSEstream';
+import { fetchStreamWithAuth } from '@/features/chatbot/utils/fetchStreamWithAuth';
+import { SessionHistoryData } from '@/features/comment/types/session';
 
 // location.state 타입 정의
 interface ILocationState {
@@ -246,54 +248,6 @@ const ChattingPage = () => {
     if (e.key === 'Enter' && !e.shiftKey) {
       e.preventDefault();
       handleSendMessage();
-    }
-  };
-
-  const refreshToken = async (): Promise<boolean> => {
-    try {
-      const response = await fetch(`${API_BASE_URL}${API_SUB_URLS}/auth/refresh`, {
-        method: 'POST',
-        credentials: 'include',
-      });
-
-      if (response.ok) {
-        return true;
-      } else {
-        // 리프레시 실패 시 로그인 페이지로 리다이렉트
-        navigate('/');
-
-        return false;
-      }
-    } catch (error) {
-      console.error('Token refresh failed:', error);
-      navigate('/');
-
-      return false;
-    }
-  };
-
-  // 스트리밍 응답 처리를 위한 401 에러 핸들링 함수
-  const fetchStreamWithAuth = async (url: string, options: RequestInit = {}): Promise<Response> => {
-    try {
-      const response = await fetch(url, options);
-
-      if (response.status === 401) {
-        // 토큰 리프레시 시도
-        const refreshSuccess = await refreshToken();
-
-        if (refreshSuccess) {
-          // 리프레시 성공 시 원래 요청 재시도
-          return await fetch(url, options);
-        } else {
-          // 리프레시 실패 시 에러 throw
-          throw new Error('Authentication failed');
-        }
-      }
-
-      return response;
-    } catch (error) {
-      console.error('Fetch with auth failed:', error);
-      throw error;
     }
   };
